[refactoring] 패키지명 step 5 추가
