+++ src/main/kotlin/step3_4_5/racing/CarRacing.kt
@@ -1,4 +1,4 @@
-package step3_4.racing
+package step3_4_5.racing
 
 class CarRacing(
     private val carList: List<Car>,
