+++ src/main/kotlin/step3_4_5/view/InputView.kt
@@ -1,6 +1,6 @@
-package step3_4.view
+package step3_4_5.view
 
-import step3_4.validator.RoundNumberValidator
+import step3_4_5.validator.RoundNumberValidator
 
 object InputView {
