+++ src/test/kotlin/step3_4/racing/car/CarTest.kt
@@ -3,7 +3,7 @@ package step3_4.racing.car
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import step3_4.racing.Car
+import step3_4_5.racing.Car
 
 class CarTest : StringSpec({
     "current_step이 기본 값을 잘 반환한다" {
